Working on functionality to allow flexible directory and output file naming based on sensitivity analysis.

+++ gain/gain_year_count_mangrove.py
@@ -41,7 +41,7 @@ def create_gain_year_count_loss_only(tile_id):
     loss_outfilename = '{}_growth_years_loss_only.tif'.format(tile_id)
     loss_outfilearg = '--outfile={}'.format(loss_outfilename)
     cmd = ['gdal_calc.py', '-A', loss, '-B', gain, '-C', mangrove, loss_calc, loss_outfilearg,
-           '--NoDataValue=0', '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'int16']
+           '--NoDataValue=0', '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'Byte']
     subprocess.check_call(cmd)
 
     # Prints information about the tile that was just processed
@@ -64,7 +64,7 @@ def create_gain_year_count_gain_only_standard(tile_id):
     gain_outfilename = '{}_growth_years_gain_only.tif'.format(tile_id)
     gain_outfilearg = '--outfile={}'.format(gain_outfilename)
     cmd = ['gdal_calc.py', '-A', loss, '-B', gain, '-C', mangrove, gain_calc, gain_outfilearg, '--NoDataValue=0',
-           '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'int16']
+           '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'Byte']
     subprocess.check_call(cmd)
 
     # Prints information about the tile that was just processed
@@ -87,7 +87,7 @@ def create_gain_year_count_gain_only_maxgain(tile_id):
     gain_outfilename = '{}_growth_years_gain_only.tif'.format(tile_id)
     gain_outfilearg = '--outfile={}'.format(gain_outfilename)
     cmd = ['gdal_calc.py', '-A', loss, '-B', gain, '-C', mangrove, gain_calc, gain_outfilearg, '--NoDataValue=0',
-           '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'int16']
+           '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'Byte']
     subprocess.check_call(cmd)
 
     # Prints information about the tile that was just processed
@@ -111,7 +111,7 @@ def create_gain_year_count_no_change(tile_id):
     no_change_outfilename = '{}_growth_years_no_change.tif'.format(tile_id)
     no_change_outfilearg = '--outfile={}'.format(no_change_outfilename)
     cmd = ['gdal_calc.py', '-A', loss, '-B', gain, '-C', mangrove, no_change_calc, no_change_outfilearg,
-           '--NoDataValue=0', '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'int16']
+           '--NoDataValue=0', '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'Byte']
     subprocess.check_call(cmd)
 
     # Prints information about the tile that was just processed
@@ -138,7 +138,7 @@ def create_gain_year_count_loss_and_gain_standard(tile_id):
     loss_and_gain_outfilename = '{}_growth_years_loss_and_gain.tif'.format(tile_id)
     loss_and_gain_outfilearg = '--outfile={}'.format(loss_and_gain_outfilename)
     cmd = ['gdal_calc.py', '-A', loss, '-B', gain, '-C', mangrove, loss_and_gain_calc, loss_and_gain_outfilearg,
-           '--NoDataValue=0', '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'int16']
+           '--NoDataValue=0', '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'Byte']
     subprocess.check_call(cmd)
 
     # Prints information about the tile that was just processed
@@ -161,7 +161,7 @@ def create_gain_year_count_loss_and_gain_maxgain(tile_id):
     loss_and_gain_outfilename = '{}_growth_years_loss_and_gain.tif'.format(tile_id)
     loss_and_gain_outfilearg = '--outfile={}'.format(loss_and_gain_outfilename)
     cmd = ['gdal_calc.py', '-A', loss, '-B', gain, '-C', mangrove, loss_and_gain_calc, loss_and_gain_outfilearg,
-           '--NoDataValue=0', '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'int16']
+           '--NoDataValue=0', '--overwrite', '--co', 'COMPRESS=LZW', '--type', 'Byte']
     subprocess.check_call(cmd)
 
     # Prints information about the tile that was just processed
@@ -185,7 +185,7 @@ def create_gain_year_count_merge(tile_id, pattern):
     print "  Merging loss, gain, no change, and loss/gain pixels into single raster"
     age_outfile = '{}_{}.tif'.format(tile_id, pattern)
     cmd = ['gdal_merge.py', '-o', age_outfile, loss_outfilename, gain_outfilename, no_change_outfilename, loss_and_gain_outfilename,
-           '-co', 'COMPRESS=LZW', '-a_nodata', '0', '-ot', 'int16']
+           '-co', 'COMPRESS=LZW', '-a_nodata', '0', '-ot', 'Byte']
     subprocess.check_call(cmd)
 
     # Prints information about the tile that was just processed
